backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Redis
redis==5.0.1
aioredis==2.0.1

# MinIO / S3
minio==7.2.4

# AI / LLM
anthropic==0.18.1

# OCR for Pokemon TCG Live screenshots
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Video processing
moviepy==1.0.3

# Data processing
pandas==2.2.0
pydantic==2.6.1
pydantic-settings==2.2.1

# HTTP client
httpx==0.27.0
aiohttp==3.9.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.2.3

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.27.0
